KDSource as a plugin (Target + Config). Header and source files were separated.
