+++ include/kdsource/KDSourceConfig.h.in
